flask/requirements.txt
aniso8601==9.0.1
click==7.1.2
Flask==1.1.2
Flask-RESTful==0.3.8
gunicorn==20.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
Pillow==8.2.0
pytz==2021.1
six==1.16.0
uWSGI==2.0.19.1
Werkzeug==1.0.1
